feature/movie_GenAI/requirements.txt
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
supabase>=1.0.3
python-dotenv>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0
moviepy>=1.0.3 